update voter info

+++ Political Preparedness/app/src/main/java/com/example/android/politicalpreparedness/database/ElectionDao.kt
@@ -22,11 +22,11 @@ interface ElectionDao {
 
     //TODO: Add select single election query
     @Query("SELECT * FROM election_table WHERE id == :electionId")
-    suspend fun getElectionById(electionId: Int) : Election
+    suspend fun getElectionById(electionId: Long) : Election
 
     //TODO: Add delete query
     @Query("DELETE FROM election_table WHERE id == :electionId")
-    suspend fun removeElection(electionId: Int)
+    suspend fun removeElection(electionId: Long)
 
     //TODO: Add clear query
     @Query("DELETE FROM election_table")
